pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecoff-fitter"
version = "0.1.0"
description = "A Python tool to fit an interval regression model to censored Minimum Inhibitory Concentration values"
readme = "README.md"
authors = [
    { name = "Dylan Adlard", email = "dylan.adlard@ndm.ox.ac.uk" },
    { name = "Philip W Fowler" }
]
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["ECOFF", "antimicrobial resistance", "minimum inhibitory concentration", "MIC"]
classifiers = [
    "Programming Language :: Python :: 3"
]
dependencies = [
    "numpy",
    "scipy",
    "pandas",
    "joblib",
    "matplotlib",
    "PyYAML"
]

[project.urls]
Homepage = "https://github.com/DylanAdlard/ECOFFitter"

[project.scripts]
ecoff-fitter = "ecoff_fitter.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.coverage.run]
omit = [
    "*/ecoff_fitter/wts.py",
]
